Added Fork Recovery Test

+++ mocknet/mocknet.py
@@ -10,6 +10,7 @@ import os
 import shutil
 import signal
 import subprocess
+from time import sleep
 from concurrent.futures import ThreadPoolExecutor
 from multiprocessing import Queue
 
@@ -31,7 +32,8 @@ class MockNet(object):
     def __init__(self,
                  test_function,
                  timeout_secs=60,
-                 node_count=0):
+                 node_count=0,
+                 remove_data=True):
         print("")
         self.writeout("Starting mocknet")
 
@@ -49,8 +51,9 @@ class MockNet(object):
 
         self.nodes_pids = Queue()
 
-        # Clear mocknet data
-        shutil.rmtree(self.data_dir, ignore_errors=True)
+        if remove_data:
+            # Clear mocknet data
+            shutil.rmtree(self.data_dir, ignore_errors=True)
 
     def prepare_source(self):
         cmd = "{}/prepare_source.sh".format(self.this_dir)
@@ -110,6 +113,7 @@ class MockNet(object):
 
             for node_idx in range(self.node_count):
                 self.nodes.append(self.pool.submit(self.start_node, node_idx, stop_event))
+                sleep(2)  # Delay before starting each node, so that nodes can connect to each other
 
             try:
                 result = test_future.result(self.timeout_secs)
